+++ packages/react-app/plans/nft1.json
@@ -1,7 +1,7 @@
 {
   "description": "Beginning of great things starts with a step",
   "image": "https://ipfs.io/ipfs/QmRTVJ7ynQU12KWPsP1oxbFBEjUaHoJdsNgWJUsbjtsXiZ/nft1.jpg",
-  "name": "Chain",
+  "name": "Helping Hand",
   "attributes": [
     {
       "trait_type": "private_forum_access",
