+++ packages/react-app/plans/nft2.json
@@ -1,7 +1,7 @@
 {
   "description": "Very human of you to be yourself in the world",
   "image": "https://ipfs.io/ipfs/QmRTVJ7ynQU12KWPsP1oxbFBEjUaHoJdsNgWJUsbjtsXiZ/nft2.jpg",
-  "name": "Chain",
+  "name": "Go Solar",
   "attributes": [
     {
       "trait_type": "private_forum_access",
