+++ packages/react-app/plans/nft3.json
@@ -1,7 +1,7 @@
 {
   "description": "Green world with a touch of kindness",
   "image": "https://ipfs.io/ipfs/QmRTVJ7ynQU12KWPsP1oxbFBEjUaHoJdsNgWJUsbjtsXiZ/nft3.jpg",
-  "name": "Chain",
+  "name": "NevaWaste",
   "attributes": [
     {
       "trait_type": "private_forum_access",
